requirements.txt
azure-storage-blob==12.9.0
Django==3.2.12
python-dotenv==0.19.1
